Rename the types folder to configs
