+++ incoming-message-handler/src/configs/DynamoDbConfig.ts
