+++ incoming-message-handler/src/configs/createDynamoDbConfig.ts
