+++ incoming-message-handler/src/configs/index.ts
@@ -2,6 +2,5 @@ export { default as DynamoDbConfig } from "./DynamoDbConfig"
 export { default as createDynamoDbConfig } from "./createDynamoDbConfig"
 export { default as MqConfig } from "./MqConfig"
 export { default as createMqConfig } from "./createMqConfig"
-export { default as MessageData } from "./MessageData"
 export { default as S3Config } from "./S3Config"
 export { default as createS3Config } from "./createS3Config"
