+++ incoming-message-handler/src/e2e/e2e..test.ts
@@ -2,7 +2,7 @@ import { v4 as uuid } from "uuid"
 import format from "xml-formatter"
 import { isError } from "@handlers/common"
 import TestDynamoGateway from "../gateways/DynamoGateway/TestDynamoGateway"
-import IncomingMessage from "../entities/IncomingMessage"
+import { IncomingMessage } from "../entities"
 import IncomingMessageSimulator from "./IncomingMessageSimulator"
 import IbmMqService from "./IbmMqService"
 import TestS3Gateway from "../gateways/S3Gateway/TestS3Gateway"
