+++ incoming-message-handler/src/entities/index.ts
@@ -1,0 +1,3 @@
+export { default as MessageData } from "./MessageData"
+export { default as DeliveryMessage } from "./DeliveryMessage"
+export { default as IncomingMessage } from "./IncomingMessage"
